+++ contracts/eosio.system/src/voting.cpp
@@ -26,6 +26,48 @@ namespace {
 
 } // namespace
 
+
+////////////////////////////////////////////////////////////////////////////////
+/// @todo This should be moved to CDT
+
+#include <variant>
+#include <vector>
+
+namespace eosio {
+   namespace internal_use_do_not_use {
+      extern "C" {
+        __attribute((eosio_wasm_import))
+        int64_t set_proposed_producers( char*, uint32_t );
+      }
+   }
+
+   using schedule_version = uint64_t;
+
+   enum class proposed_producers_errors {
+      disallow_empty_producer_schedule = -1,
+      already_proposed_schedule = -2,
+      schedule_does_not_change = -3,
+      schedule_would_not_change = -4
+   };
+
+   auto set_proposed_producers_ex(const std::vector<producer_key>& prods)
+      -> std::variant<schedule_version, proposed_producers_errors>
+   {
+      auto packed_prods = eosio::pack(prods);
+
+      int64_t ret = internal_use_do_not_use::set_proposed_producers(
+         static_cast<char*>(packed_prods.data()), packed_prods.size());
+
+      if (ret >= 0)
+         return static_cast<schedule_version>(ret);
+      else
+         return static_cast<proposed_producers_errors>(ret);
+   }
+} // namespace eosio
+
+// End of CDT section
+////////////////////////////////////////////////////////////////////////////////
+
 namespace eosiosystem {
 
    using eosio::const_mem_fun;
@@ -61,6 +103,7 @@ namespace eosiosystem {
             update_total_votepay_share( ct, 0.0, prod->total_votes );
             // When introducing the producer2 table row for the first time, the producer's votes must also be accounted for in the global total_producer_votepay_share at the same time.
          }
+
       } else {
          _producers.emplace( producer, [&]( producer_info& info ){
             info.owner           = producer;
@@ -71,10 +114,23 @@ namespace eosiosystem {
             info.location        = location;
             info.last_claim_time = ct;
          });
+
          _producers2.emplace( producer, [&]( producer_info2& info ){
             info.owner                     = producer;
             info.last_votepay_share_update = ct;
          });
+
+         if (_grewards.activated) {
+            _rewards.emplace( producer, [&]( rewards_info& info ){
+               info.init(producer);
+
+               // If we only have 21 producers or less they are ready to produce, otherwise
+               // they will have to wait to be selected 
+               /// @todo It's necessary to check for "active" producers.
+               if (std::distance(_producers.cbegin(), _producers.cend()) < 21) 
+                  info.select(reward_type::producer);
+            });
+         }
       }
 
    }
@@ -88,7 +144,51 @@ namespace eosiosystem {
       });
    }
 
-   void system_contract::select_producers_into( uint64_t begin, uint64_t count, prod_vec_t& result ) {
+   /**
+    * Updates the reward status of all producers (including those that are not top producers)
+    */
+   void system_contract::update_producer_reward_status(const prod_vec_t& top_producers) {
+      auto idx = _producers.get_index<"prototalvote"_n>();
+      uint64_t i = 0;
+
+      for (auto it = idx.cbegin(); it != idx.cend() && it->active(); ++it, ++i) {
+         // Look for the status in the top producer list
+         auto prod_it = std::lower_bound(
+            top_producers.begin(), 
+            top_producers.end(), 
+            it->owner, 
+            [](const auto& prod_tuple, const auto& owner) -> bool { 
+               return std::get<0>(prod_tuple).producer_name < owner; 
+            });
+
+         if (prod_it != top_producers.end()) {
+            const auto& [ prod_key, _, type ] = *prod_it;
+            
+            if (auto reward_it = _rewards.find(prod_key.producer_name.value); reward_it != _rewards.end()) {
+               // type = type => workaround for a limitation of capturing structured bindings
+               _rewards.modify(reward_it, same_payer, [type = type](auto& rec) { 
+                  rec.select(type);
+               });
+            }
+         }
+         else {
+            if (auto reward_it = _rewards.find(it->owner.value); reward_it != _rewards.end()) {
+               _rewards.modify(reward_it, same_payer, [&](auto& rec) {
+                  rec.set_current_type(reward_type::none);
+               });
+            }
+         }
+      } 
+   }
+
+   /**
+    * Selects the specified producer range into the result vector. It also adds the
+    * provided status to that vector. 
+    */
+   void system_contract::select_producers_into( uint64_t begin, 
+                                                uint64_t count,
+                                                reward_type type, 
+                                                prod_vec_t& result ) {
       auto idx = _producers.get_index<"prototalvote"_n>();
       uint64_t i = 0;
 
@@ -98,7 +198,7 @@ namespace eosiosystem {
       {
          if (i >= begin)
             result.emplace_back(
-               prod_vec_t::value_type{{it->owner, it->producer_key}, it->location});
+               prod_vec_t::value_type{{it->owner, it->producer_key}, it->location, type});
       }
    }
 
@@ -108,7 +208,6 @@ namespace eosiosystem {
 
       auto constexpr total_weight = 1'000'000;
       auto constexpr one_percent_weight = total_weight * 0.01;
-      auto constexpr num_standbys = 36;
       auto constexpr standby_weight = 10 * one_percent_weight / num_standbys;
 
       auto const selected_weight = to_int(previous_block_hash) % total_weight;
@@ -120,19 +219,20 @@ namespace eosiosystem {
       if (standby_index < num_standbys) {
          prod_vec_t standbys;
          standbys.reserve(num_standbys);
-         select_producers_into(21, num_standbys, standbys);
+         select_producers_into(21, num_standbys, reward_type::standby, standbys);
 
          if (standbys.size() > standby_index) {
             top_producers.emplace_back(standbys[standby_index]);
 
             eosio::print_f("Standby producer '%' was selected\n", 
-               standbys[standby_index].first.producer_name.to_string());
+               std::get<0>(standbys[standby_index]).producer_name.to_string());
          }
       }
 
-      select_producers_into(0, 21 - top_producers.size(), top_producers);
+      select_producers_into(0, 21 - top_producers.size(), reward_type::producer, top_producers);
 
       if (top_producers.size() == 0 || top_producers.size() < _gstate.last_producer_schedule_size ) {
+         eosio::print("No top producers or they are less than the last scheduled");
          return;
       }
 
@@ -143,10 +243,27 @@ namespace eosiosystem {
 
       producers.reserve(top_producers.size());
       for( const auto& item : top_producers )
-         producers.push_back(item.first);
+         producers.push_back(std::get<0>(item));
 
-      if( set_proposed_producers( producers ) >= 0 ) {
+      auto result = set_proposed_producers_ex( producers );
+
+      using eosio::proposed_producers_errors;
+
+      if (auto err = std::get_if<proposed_producers_errors>(&result)) {
+         switch(*err) {
+            case proposed_producers_errors::schedule_does_not_change:
+            case proposed_producers_errors::schedule_would_not_change:
+               update_producer_reward_status(top_producers);
+               break;
+
+            default:
+               ;
+         }
+      }
+      else {
+         // New list accepted
          _gstate.last_producer_schedule_size = static_cast<decltype(_gstate.last_producer_schedule_size)>( top_producers.size() );
+         update_producer_reward_status(top_producers);
       }
    }